requirements.txt
feedparser>=6.0
youtube-transcript-api>=1.2
google-genai>=1.0
requests>=2.31
pytest>=7.0
